debut du responcive design + verification du contenue de l'objet contact pour bloquer si vide

+++ frontend/ressources/js/panier.js
@@ -30,21 +30,15 @@ createPanier = () => {
 
         tablePanier.forEach(panier => {
 
-
             const tableauProduit = document.getElementById('panier_produit');
 
             const newTr = tableauProduit.appendChild(createElm1('tr', '', 'id', 'bloc-tr'));
 
-            // key provisoir a supprimer quand suppression produit fini
-            newTr.appendChild(createElm1('td', panier.key, 'class', ''));
-            //fin
-
-            newTr.appendChild(createElm1('td', panier.name, 'class', 'bloc__table--align'));
-            newTr.appendChild(createElm1('td', '1', 'class', 'bloc__table--align'));
-            newTr.appendChild(createElm1('td', pricesProduct(panier), 'class', 'bloc__table--align'));
-            newTr.appendChild(createElm1('td', pricesProduct(panier), 'class', 'bloc__table--align'));
-            newTr.appendChild(createElm1('td', '', 'class', 'bloc__table--align'));
-            newTr.appendChild(createElm1('button', 'Supprimer', 'class', 'bloc__table__button--style btnDeleteO'));
+            newTr.appendChild(createElm1('td', panier.name, 'class', 'bloc__td--border bloc__td--style2 bloc--table--align--font'));
+            newTr.appendChild(createElm1('td', '1', 'class', 'bloc__td--style2'));
+            newTr.appendChild(createElm1('td', pricesProduct(panier), 'class', 'bloc__td--style2'));
+            newTr.appendChild(createElm1('td', pricesProduct(panier), 'class', 'bloc__td--style2'));
+            newTr.appendChild(createElm1('td', 'Supprimer', 'class', 'bloc__td--style2 bloc__table__button--style2 btnDeleteO'));
 
             //-----suppression produit------
             newTr.addEventListener('click', function(i) {
@@ -57,12 +51,16 @@ createPanier = () => {
         const totalProduit = document.getElementById('total_panier');
 
         const newTrTotal = document.createElement('tr');
+        newTrTotal.setAttribute('class', '');
         totalProduit.appendChild(newTrTotal);
 
         // prix total du panier
-        newTrTotal.appendChild(createElm1('td', totalpanier(), 'id', 'totalCommande'));
+        newTrTotal.appendChild(createElm1('td', 'total du panier :', 'class', ' bloc__table--padding--colors--price--total1'));
+        newTrTotal.appendChild(createElm2('td', totalpanier(), 'id', 'totalCommande', 'class', ' bloc__table--padding--colors--price--total2'));
+        newTrTotal.appendChild(createElm2('td', '', 'id', 'btn2', 'class', ''));
 
-        newTrTotal.appendChild(createElm2('button', 'tous supprimer', 'class', 'bloc__table__button--style', 'id', 'deleteTotalPanier'));
+        const newbtn2 = document.getElementById('btn2');
+        newbtn2.appendChild(createElm2('button', 'tous supprimer', 'class', 'bloc__table__button--style', 'id', 'deleteTotalPanier'));
 
         document.getElementById('deleteTotalPanier').addEventListener('click', deleteTotalPanier => {
             localStorage.clear();
@@ -72,39 +70,46 @@ createPanier = () => {
         // formulaire pour commande
         const formClient = document.getElementById('formulaire');
 
-        const newdiv1 = formClient.appendChild(createElm1('div', '', 'class', 'bloc__form--flex2'));
+        const newdiv0 = formClient.appendChild(createElm1('div', '', 'class', 'bloc__form__div__width'));
 
-        newdiv1.appendChild(createElm2('label', 'Prenom :', 'class', 'bloc__form__label--font', 'for', 'prenom'));
-        newdiv1.appendChild(createElm2('label', 'Nom :', 'class', 'bloc__form__label--font1', 'for', 'nom'));
-        newdiv1.appendChild(createElm2('label', 'Adresse :', 'class', 'bloc__form__label--font1', 'for', 'adresse'));
-        newdiv1.appendChild(createElm2('label', 'Ville :', 'class', 'bloc__form__label--font1', 'for', 'ville'));
-        newdiv1.appendChild(createElm2('label', 'Adresse électronique :', 'class', 'bloc__form__label--font1', 'for', 'email'));
+        newdiv0.appendChild(createElm2('p', '* champ obligatoire', 'id', 'erreur1', 'class', 'bloc__from--p--flex'));
+        const newdiv1 = formClient.appendChild(createElm1('div', '', 'class', 'bloc__form__flex--input--label'));
+        newdiv1.appendChild(createElm2('label', 'Prenom :', 'class', 'bloc__form__label--font--seize', 'for', 'prenom'));
+        newdiv1.appendChild(createinputs('input', 'id', 'prenom', 'name', 'prenom', 'class', 'bloc__form__input--border', 'type', 'text', 'value', ''));
+        newdiv0.appendChild(newdiv1);
 
-
-        const newdiv2 = formClient.appendChild(createElm1('div', '', 'class', 'bloc__form--flex2 bloc__form--padding'));
-
-        newdiv2.appendChild(createElm1('p', '* champ obligatoire', 'id', 'erreur1'));
-        newdiv2.appendChild(createinputs('input', 'id', 'prenom', 'name', 'prenom', 'class', 'bloc__form__input--border', 'type', 'text', 'value', ''));
-
-        newdiv2.appendChild(createElm1('p', '* champ obligatoire', 'id', 'erreur2'));
+        newdiv0.appendChild(createElm2('p', '* champ obligatoire', 'id', 'erreur2', 'class', 'bloc__from--p--flex'));
+        const newdiv2 = formClient.appendChild(createElm1('div', '', 'class', 'bloc__form__flex--input--label'));
+        newdiv2.appendChild(createElm2('label', 'Nom :', 'class', 'bloc__form__label--font--seize', 'for', 'nom'));
         newdiv2.appendChild(createinputs('input', 'id', 'nom', 'name', 'nom', 'class', 'bloc__form__input--border', 'type', 'text', 'value', ''));
+        newdiv0.appendChild(newdiv2);
 
-        newdiv2.appendChild(createElm1('p', '* champ obligatoire', 'id', 'erreur3'));
-        newdiv2.appendChild(createinputs('input', 'id', 'adresse', 'name', 'adresse', 'class', 'bloc__form__input--border', 'type', 'text', 'value', ''));
+        newdiv0.appendChild(createElm2('p', '* champ obligatoire', 'id', 'erreur3', 'class', 'bloc__from--p--flex'));
+        const newdiv3 = formClient.appendChild(createElm1('div', '', 'class', 'bloc__form__flex--input--label'));
+        newdiv3.appendChild(createElm2('label', 'Adresse :', 'class', 'bloc__form__label--font--seize', 'for', 'adresse'));
+        newdiv3.appendChild(createinputs('input', 'id', 'adresse', 'name', 'adresse', 'class', 'bloc__form__input--border', 'type', 'text', 'value', ''));
+        newdiv0.appendChild(newdiv3);
 
-        newdiv2.appendChild(createElm1('p', '* champ obligatoire', 'id', 'erreur4'));
-        newdiv2.appendChild(createinputs('input', 'id', 'ville', 'name', 'ville', 'class', 'bloc__form__input--border', 'type', 'text', 'value', ''));
+        newdiv0.appendChild(createElm2('p', '* champ obligatoire', 'id', 'erreur4', 'class', 'bloc__from--p--flex'));
+        const newdiv4 = formClient.appendChild(createElm1('div', '', 'class', 'bloc__form__flex--input--label'));
+        newdiv4.appendChild(createElm2('label', 'Ville :', 'class', 'bloc__form__label--font--seize', 'for', 'ville'));
+        newdiv4.appendChild(createinputs('input', 'id', 'ville', 'name', 'ville', 'class', 'bloc__form__input--border', 'type', 'text', 'value', ''));
+        newdiv0.appendChild(newdiv4);
 
-        newdiv2.appendChild(createElm1('p', '* champ obligatoire', 'id', 'erreur5'));
-        newdiv2.appendChild(createinputs('input', 'id', 'email', 'name', 'email', 'class', 'bloc__form__input--border', 'type', 'email', 'value', ''));
+        newdiv0.appendChild(createElm2('p', '* champ obligatoire', 'id', 'erreur5', 'class', 'bloc__from--p--flex'));
+        const newdiv5 = formClient.appendChild(createElm1('div', '', 'class', 'bloc__form__flex--input--label'));
+        newdiv5.appendChild(createElm2('label', 'Adresse électronique :', 'class', 'bloc__form__label--font--seize', 'for', 'email'));
+        newdiv5.appendChild(createinputs('input', 'id', 'email', 'name', 'email', 'class', 'bloc__form__input--border', 'type', 'email', 'value', ''));
+        newdiv0.appendChild(newdiv5);
 
-        newdiv2.appendChild(createElm3('button', 'Commander', 'id', 'envoyer_commande', 'class', 'bloc__section_4__button--style', 'type', 'submit'));
+        newdiv0.appendChild(createElm3('div', 'Commander', 'id', 'envoyer_commande', 'class', ' bloc__section_4__button--style', 'type', 'submit'));
 
     } //fin de create panier
 
 checkPanier();
 
 // button commander + objet contact
+
 validerForms = () => {
 
     const product_id = [];
@@ -127,56 +132,52 @@ validerForms = () => {
     const regexVille = /^[a-zA-Za-zA-Z0-9áàâäãåçéèêëíìîïñóòôöõúùûüýÿæœÁÀÂÄÃÅÇÉÈÊËÍÌÎÏÑÓÒÔÖÕÚÙÛÜÝŸÆŒ\-]+/;
     const regexEmail = /^[a-zA-Z1-9\-]+?@{1}[a-zA-Z1-9]+[.]{1}[a-zA-Z1-9]+/;
 
+    btnEnvoi.addEventListener('click', function() {
 
-    btnEnvoi.addEventListener('click', validation);
+        function verif() {
 
-    function validation(event, i) {
+            if (regexNomPrenom.test(prenom.value) == true) {
+                paragErreur1.setAttribute('class', 'bloc__form--font--erreur2 bloc__from--p--flex');
+                paragErreur1.innerHTML = '* champ obligatoire';
+            } else if (regexNomPrenom.test(prenom.value) == false) {
+                paragErreur1.setAttribute('class', 'bloc__form--font--erreur bloc__from--p--flex');
+                paragErreur1.innerHTML = 'Format du PRENOM non  conforme !!!';
+            };
 
-        event.preventDefault();
+            if (regexNomPrenom.test(nom.value) == true) {
+                paragErreur2.setAttribute('class', 'bloc__form--font--erreur2 bloc__from--p--flex');
+                paragErreur2.innerHTML = '* champ obligatoire';
+            } else if (regexNomPrenom.test(nom.value) == false) {
+                paragErreur2.setAttribute('class', 'bloc__form--font--erreur bloc__from--p--flex');
+                paragErreur2.innerHTML = 'Format du NOM non conforme !!!';
+            };
 
-        if (regexNomPrenom.test(prenom.value) == true) {
-            //console.log(typeof prenom.value);
-            paragErreur1.setAttribute('class', 'bloc__form--font--erreur2');
-            paragErreur1.innerHTML = '* champ obligatoire';
-        } else if (regexNomPrenom.test(prenom.value) == false) {
-            paragErreur1.setAttribute('class', 'bloc__form--font--erreur');
-            paragErreur1.innerHTML = 'Format du PRENOM non  conforme !!!';
-        };
+            if (regexAdresse.test(adresse.value) == true) {
+                paragErreur3.setAttribute('class', 'bloc__form--font--erreur2 bloc__from--p--flex');
+                paragErreur3.innerHTML = '* champ obligatoire';
+            } else if (regexAdresse.test(adresse.value) == false) {
+                paragErreur3.setAttribute('class', 'bloc__form--font--erreur bloc__from--p--flex');
+                paragErreur3.innerHTML = 'Format de l\'ADRESSE non conforme !!!';
+            };
 
-        if (regexNomPrenom.test(nom.value) == true) {
-            //console.log(typeof nom.value);
-            paragErreur2.setAttribute('class', 'bloc__form--font--erreur2');
-            paragErreur2.innerHTML = '* champ obligatoire';
-        } else if (regexNomPrenom.test(nom.value) == false) {
-            paragErreur2.setAttribute('class', 'bloc__form--font--erreur');
-            paragErreur2.innerHTML = 'Format du NOM non conforme !!!';
-        };
+            if (regexVille.test(ville.value) == true) {
+                paragErreur4.setAttribute('class', 'bloc__form--font--erreur2 bloc__from--p--flex');
+                paragErreur4.innerHTML = '* champ obligatoire';
+            } else if (regexVille.test(ville.value) == false) {
+                paragErreur4.setAttribute('class', 'bloc__form--font--erreur bloc__from--p--flex');
+                paragErreur4.innerHTML = 'Format de la VILLE non conforme !!!';
+            };
 
-        if (regexAdresse.test(adresse.value) == true) {
-            //console.log(typeof adresse.value);
-            paragErreur3.setAttribute('class', 'bloc__form--font--erreur2');
-            paragErreur3.innerHTML = '* champ obligatoire';
-        } else if (regexAdresse.test(adresse.value) == false) {
-            paragErreur3.setAttribute('class', 'bloc__form--font--erreur');
-            paragErreur3.innerHTML = 'Format de l\'ADRESSE non conforme !!!';
-        };
-        if (regexVille.test(ville.value) == true) {
-            //console.log(typeof ville.value);
-            paragErreur4.setAttribute('class', 'bloc__form--font--erreur2');
-            paragErreur4.innerHTML = '* champ obligatoire';
-        } else if (regexVille.test(ville.value) == false) {
-            paragErreur4.setAttribute('class', 'bloc__form--font--erreur');
-            paragErreur4.innerHTML = 'Format de la VILLE non conforme !!!';
-        };
+            if (regexEmail.test(email.value) == true) {
+                paragErreur5.setAttribute('class', 'bloc__form--font--erreur2 bloc__from--p--flex');
+                paragErreur5.innerHTML = '* champ obligatoire';
+            } else if (regexEmail.test(email.value) == false) {
+                paragErreur5.setAttribute('class', 'bloc__form--font--erreur bloc__from--p--flex');
+                paragErreur5.innerHTML = 'Format de l\'E-MAIL non conforme !!!';
+            };
+        }; //fin verif
 
-        if (regexEmail.test(email.value) == true) {
-            //console.log(typeof email.value);
-            paragErreur5.setAttribute('class', 'bloc__form--font--erreur2');
-            paragErreur5.innerHTML = '* champ obligatoire';
-        } else if (regexEmail.test(email.value) == false) {
-            paragErreur5.setAttribute('class', 'bloc__form--font--erreur');
-            paragErreur5.innerHTML = 'Format de l\'E-MAIL non conforme !!!';
-        };
+        verif();
 
         const contact = {
             firstName: prenom.value,
@@ -186,8 +187,6 @@ validerForms = () => {
             email: email.value
         };
 
-
-
         const storage2 = localStorage.length;
 
         for (let i = 0; i < storage2; i++) {
@@ -196,27 +195,36 @@ validerForms = () => {
 
             product_id.push(objJsonParse2.id);
         };
+
         const order = {
             contact: contact,
             products: product_id
         };
-        const datas = send('http://localhost:3000/api/teddies/order', order);
-        console.log(order);
-        datas.then(rep => {
 
-            localStorage.clear();
+        if (contact.firstName == false, contact.lastName == false, contact.address == false, contact.ville == false, contact.email == false) {
+            alert('formulaire mal rempli');
+        } else {
+            const datas = send('http://localhost:3000/api/teddies/order', order);
+            datas.then(rep => {
+                console.log(rep);
 
-            const repstring = JSON.stringify(rep);
-            localStorage.setItem('repOrder', repstring);
+                localStorage.clear();
 
-            window.location = 'orinoco_confirmation_commande.html';
+                const repstring = JSON.stringify(rep);
+                localStorage.setItem('repOrder', repstring);
+                console.log(localStorage);
 
-            //console.log(rep);
-            //console.log(localStorage);
-            // console.log(rep.contact.email);
-        });
+                window.location = 'orinoco_confirmation_commande.html';
 
-    };
+            }).catch((error => {
+                alert('erreur serveur');
+            })); //fin catch
+        }; // fin de else
+    }); //fin de listner
 
-};
+
+}; //fin validform
+
+
+
 validerForms();
